requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==1.2.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==3.2.8
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
annotated-types==0.6.0
paho-mqtt==2.0.0
pydantic==2.5.2
pydantic_core==2.14.5
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
typing_extensions==4.8.0
